requirements.txt
pypdf2>=3.0.1
pdfplumber>=0.10.0
ollama>=0.1.7
python-dotenv>=1.0.0
nltk>=3.8.1
scikit-learn>=1.3.2
numpy>=1.24.3
pandas>=1.5.3
rich>=13.7.0
click>=8.1.7
requests>=2.31.0
psycopg2-binary>=2.9.7

# Dev dependencies
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0